Add a --dry-run flag for provisioning
The --dry-run flag is for simulating changes to
compose. If a deployment already exists its connection
strings will be returned, but no changes to scaling,
database version or team roles will get applied. If a
deployment does not exist it will not be created, and
a fake connection string will be returned for testing
purposes.