Add endpoint_maps to the configuration input
With Compose.io enterprise accounts its possible to specify a
schema for the endpoints of deployments on a cluster to return,
but the deployments themselves may have additional endpoints
not returned by that schema, so supporting the ability to map
the values returned by Compose to the values used by a Compose
client is very useful.

This new config object is the first v2 object of several planned,
and the intermingling of v1 and v2 object is of course supported.