Break apart the connection package
Most of the stateful coordination from the connection
package has been moved into the config package and the
actual running of pachelbel actions (creates, updates,
lookups) is in the runner package.

Changes of this large refactoring that impact end users
1\ The 'deprovision' object
Before the `deprovision` command was implemented by
spooling up connection runners in memory in a way that
bypassed the config package and was incompatible with
the `provision` schema. The new deprovision schema
(documented in schema/v2.md) is fully compatible with
a provision call so a "provision" can now make
destructive changes.

`pachelbel deprovision` behaves mostly the same as it
did before, though it now only accepts deployment names
as arguments, not deployment IDs.

2\ Better status messages at the expense of speed
Additional pre-processing is done while validating
provisioning inputs, so we should be able to give
better messages than just 'Updating' or 'Creating',
but this pre-processing is done serially.